+++ backend/core/management/commands/load_data.py
@@ -20,6 +20,7 @@ from products.models import (
     Subcategory,
     Tag,
 )
+from recipes.models import ProductsInRecipe, Recipe
 from reviews.models import Review
 from users.models import Address, User
 
@@ -288,6 +289,7 @@ def read_reviews():
                 pub_date=row["pub_date"],
                 author_id=row["author_id"],
                 product_id=row["product_id"],
+                was_edited=row["was_edited"],
             )
             review.save()
 
@@ -302,6 +304,41 @@ def read_sessions():
                 expire_date=row["expire_date"],
             )
             session.save()
+
+
+def read_recipes():
+    with open(os.path.join(DATA_DIR, "recipes.csv"), "r", encoding="utf-8") as f:
+        reader = DictReader(f)
+        for row in reader:
+            if row.get("pub_date"):
+                pub_date = row["pub_date"]
+            else:
+                pub_date = timezone.now()
+            recipe = Recipe(
+                id=row["id"],
+                pub_date=pub_date,
+                author_id=row["author_id"],
+                name=row["name"],
+                image=row["image"],
+                text=row["text"],
+                cooking_time=row["cooking_time"],
+            )
+            recipe.save()
+
+
+def read_products_in_recipes():
+    with open(
+        os.path.join(DATA_DIR, "products_in_recipe.csv"), "r", encoding="utf-8"
+    ) as f:
+        reader = DictReader(f)
+        for row in reader:
+            recipe = ProductsInRecipe(
+                id=row["id"],
+                recipe_id=row["recipe_id"],
+                ingredient_id=row["ingredient_id"],
+                amount=row["amount"],
+            )
+            recipe.save()
 
 
 class Command(BaseCommand):
@@ -344,6 +381,10 @@ class Command(BaseCommand):
         self.stdout.write("Данные из файла reviews.csv загружены")
         read_sessions()
         self.stdout.write("Данные из файла sessions.csv загружены")
+        read_recipes()
+        self.stdout.write("Данные из файла recipes.csv загружены")
+        read_products_in_recipes()
+        self.stdout.write("Данные из файла products_in_recipe.csv загружены")
 
         model_list = [
             Delivery,
@@ -362,6 +403,8 @@ class Command(BaseCommand):
             User,
             Token,
             Session,
+            Recipe,
+            ProductsInRecipe,
         ]
         sequence_sql = connection.ops.sequence_reset_sql(no_style(), model_list)
         with connection.cursor() as cursor:
